W10DC: Use the $_disks Var to list all Disks
This change adds a hard-named varoable $_disks that can be defined
by the user as a list of all disk letters that will be looked at.
The list is comma-separated, and it ignored spaces. It expects to
receive one-character values, and will reject anything else. This
is currently used in the seeker mode only, but the plan is to make
it usable in every template. Stay tuned.

In the case that the variable isn't defined, the W10DC will use
C:\ as the entry point for the search.